directory enhancement: moved all frontend scripts to frontend folder for better organization, created new directory structure for backend scripts
