+++ team/m3/js/frontend/MovingToken.js
