+++ team/m3/main.html
@@ -218,15 +218,15 @@
     </div>
     <span class="menuOpen" onclick="openNav()">&#9876;</span>
     
-    <script type="module" src="js/grid.js"></script>
+    <script type="module" src="js/frontend/grid.js"></script>
 
-    <script src="js/dice.js"></script>   <!--added this for dice roller -->
+    <script src="js/frontend/dice.js"></script>   <!--added this for dice roller -->
 
-    <script src="js/paint.js"></script>  <!--this is for the paint -->
+    <script src="js/frontend/paint.js"></script>  <!--this is for the paint -->
 
-    <script type="module" src="/team/m3/js/TileLogic.js"></script> <!-- added this for the tile feature -->
+    <script type="module" src="/team/m3/js/frontend/TileLogic.js"></script> <!-- added this for the tile feature -->
 
-    <script type="module" src="js/ObjectTokenFinal.js"></script>
+    <script type="module" src="js/frontend/ObjectTokenFinal.js"></script>
 
 
 </body>
